# Optimize CheckLength method