BinaryTokenStreamReader's <code>CheckLength</code> method takes ~21% of deserialization time (tested on simple POCO) due to excessive usage of <code>ArraySegment</code>'s properties: <code>Offset</code>, <code>Array</code> and <code>Count</code>.

![dottraceview64_2016-07-30_00-11-32](https://cloud.githubusercontent.com/assets/5787619/17269302/5cba5ad0-564d-11e6-9bca-0bb27c09644c.png)

During rewrite was removed repeatable accessing of the properties and unnecessary construction of the <code>ArraySegment</code> on each call. 

Results after:

![dottraceview64_2016-07-30_00-06-32](https://cloud.githubusercontent.com/assets/5787619/17269307/71eec648-564d-11e6-8f2e-f40ea9cab58b.png)