@dVakulen 2016-07-31
Results from https://github.com/dotnet/orleans/pull/2003:
Before: 

                Method | Repeats |      Median |    StdDev |
---------------------- |-------- |------------ |---------- |
   SerializerBenchmark |  300000 | 581.5529 ms | 2.2370 ms |
 DeserializerBenchmark |  300000 | 466.5629 ms | 1.4196 ms |

After:

                Method | Repeats |      Median |    StdDev |
---------------------- |-------- |------------ |---------- |
   SerializerBenchmark |  300000 | 578.9851 ms | 3.9789 ms |
 DeserializerBenchmark |  300000 | 403.7003 ms | 1.2128 ms |