@gabikliot 2016-08-02
So almost no improvement fro serialization and about 15% for deserialization?